+++ programmable_video/ios/Classes/PluginHandler.swift
@@ -27,6 +27,8 @@ public class PluginHandler: BaseListener {
             setSpeakerphoneOn(call, result: result)
         case "getSpeakerphoneOn":
             getSpeakerphoneOn(result: result)
+        case "deviceHasReceiver":
+            deviceHasReceiver(result: result)
         case "LocalAudioTrack#enable":
             localAudioTrackEnable(call, result: result)
         case "LocalDataTrack#sendString":
@@ -294,6 +296,33 @@ public class PluginHandler: BaseListener {
         let speakerPhoneOn = AVAudioSession.sharedInstance().mode == .videoChat
         SwiftTwilioProgrammableVideoPlugin.debug("PluginHandler.getSpeakerphoneOn => called \(speakerPhoneOn)")
         return result(speakerPhoneOn)
+    }
+
+    private func deviceHasReceiver(result: @escaping FlutterResult) {
+        let currentMode = AVAudioSession.sharedInstance().mode
+        var hasReceiver = false
+
+        if currentMode != .voiceChat {
+            do {
+                try AVAudioSession.sharedInstance().setMode(.voiceChat)
+            } catch let error as NSError {
+                return result(FlutterError(code: "\(error.code)", message: error.description, details: nil))
+            }
+        }
+
+        let receivers = AVAudioSession.sharedInstance().currentRoute.outputs
+        hasReceiver = receivers.first( where: { $0.portType == AVAudioSession.Port.builtInReceiver }) != nil
+
+        if AVAudioSession.sharedInstance().mode != currentMode {
+            do {
+                try AVAudioSession.sharedInstance().setMode(currentMode)
+            } catch let error as NSError {
+                return result(FlutterError(code: "\(error.code)", message: error.description, details: nil))
+            }
+        }
+
+        SwiftTwilioProgrammableVideoPlugin.debug("PluginHandler.deviceHasReceiver => called \(hasReceiver)")
+        return result(hasReceiver)
     }
 
     private func disconnect(_ call: FlutterMethodCall, result: @escaping FlutterResult) {