+++ programmable_video/test/programmable_video_test.dart
@@ -47,6 +47,17 @@ void main() {
     });
   });
 
+  group('.deviceHasReceiver()', () {
+    final mockInterface = MockInterface();
+    setUpAll(() => ProgrammableVideoPlatform.instance = mockInterface);
+
+    test('should call interface code to check if device has a receiver', () async {
+      final result = await TwilioProgrammableVideo.deviceHasReceiver();
+      expect(mockInterface.deviceHasReceiverWasCalled, true);
+      expect(result, true);
+    });
+  });
+
   group('.requestPermissionForCameraAndMicrophone()', () {
     test('should request camera and microphone permission', () async {
       var nativeRequestPermissionsIsCalled = false;