+++ programmable_video_platform_interface/lib/src/models/local_participant/local_participant_model.dart
@@ -1,3 +1,4 @@
+import 'package:dartlin/dartlin.dart';
 import 'package:enum_to_string/enum_to_string.dart';
 import 'package:flutter/foundation.dart';
 import 'package:twilio_programmable_video_platform_interface/src/enums/enum_exports.dart';
@@ -59,7 +60,7 @@ class LocalParticipantModel {
       }
     }
 
-    var networkQualityLevel = EnumToString.fromString(NetworkQualityLevel.values, map['networkQualityLevel']) ?? NetworkQualityLevel.NETWORK_QUALITY_LEVEL_UNKNOWN;
+    final networkQualityLevel = (map['networkQualityLevel'] as String)?.let((it) => EnumToString.fromString(NetworkQualityLevel.values, it)) ?? NetworkQualityLevel.NETWORK_QUALITY_LEVEL_UNKNOWN;
 
     return LocalParticipantModel(
       identity: map['identity'],