+++ programmable_video_platform_interface/lib/src/programmable_video_platform_interface.dart
@@ -63,6 +63,10 @@ abstract class ProgrammableVideoPlatform extends PlatformInterface {
     throw UnimplementedError('getSpeakerphoneOn() has not been implemented.');
   }
 
+  Future<bool> deviceHasReceiver() {
+    throw UnimplementedError('deviceHasReceiver() has not been implemented.');
+  }
+
   /// Calls native code to connect to a room.
   Future<int> connectToRoom(ConnectOptionsModel connectOptions) {
     throw UnimplementedError('connectToRoom() has not been implemented.');